move combobox functionality to shared hooks
